refactor : util -> uistate 패키지이동
