+++ app/src/main/java/com/recodream_aos/recordream/presentation/record/recording/RecordBottomSheetFragment.kt
@@ -15,17 +15,17 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.recodream_aos.recordream.R
 import com.recodream_aos.recordream.databinding.FragmentRecordBottomSheetBinding
 import com.recodream_aos.recordream.presentation.record.RecordViewModel
-import com.recodream_aos.recordream.util.recorder.PlayButton
-import com.recodream_aos.recordream.util.recorder.PlayButtonState
-import com.recodream_aos.recordream.util.recorder.PlayButtonState.RECORDER_PLAY
-import com.recodream_aos.recordream.util.recorder.PlayButtonState.RECORDER_STOP
-import com.recodream_aos.recordream.util.recorder.RecordButton
-import com.recodream_aos.recordream.util.recorder.RecordButtonState
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.AFTER_RECORDING
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.BEFORE_RECORDING
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.ON_RECORDING
-import com.recodream_aos.recordream.util.recorder.Recorder
-import com.recodream_aos.recordream.util.recorder.TimeStampTextView
+import com.recodream_aos.recordream.presentation.record.recording.uistate.PlayButton
+import com.recodream_aos.recordream.presentation.record.recording.uistate.PlayButtonState
+import com.recodream_aos.recordream.presentation.record.recording.uistate.PlayButtonState.RECORDER_PLAY
+import com.recodream_aos.recordream.presentation.record.recording.uistate.PlayButtonState.RECORDER_STOP
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButton
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.AFTER_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.BEFORE_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.ON_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.Recorder
+import com.recodream_aos.recordream.presentation.record.recording.uistate.TimeStampTextView
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
