+++ app/src/main/java/com/recodream_aos/recordream/presentation/record/recording/uistate/PlayButtonState.kt
@@ -1,0 +1,6 @@
+package com.recodream_aos.recordream.presentation.record.recording.uistate // ktlint-disable package-name
+
+enum class PlayButtonState {
+    RECORDER_PLAY,
+    RECORDER_STOP,
+}
