+++ app/src/main/java/com/recodream_aos/recordream/presentation/record/recording/uistate/RecordButton.kt
@@ -1,12 +1,12 @@
-package com.recodream_aos.recordream.util.recorder // ktlint-disable package-name
+package com.recodream_aos.recordream.presentation.record.recording.uistate // ktlint-disable package-name
 
 import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatImageButton
 import com.recodream_aos.recordream.R
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.AFTER_RECORDING
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.BEFORE_RECORDING
-import com.recodream_aos.recordream.util.recorder.RecordButtonState.ON_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.AFTER_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.BEFORE_RECORDING
+import com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButtonState.ON_RECORDING
 
 class RecordButton(
     context: Context,
