+++ app/src/main/java/com/recodream_aos/recordream/presentation/record/recording/uistate/RecordButtonState.kt
@@ -1,0 +1,7 @@
+package com.recodream_aos.recordream.presentation.record.recording.uistate // ktlint-disable package-name
+
+enum class RecordButtonState {
+    BEFORE_RECORDING,
+    ON_RECORDING,
+    AFTER_RECORDING,
+}
