--- app/src/main/java/com/recodream_aos/recordream/util/recorder/PlayButtonState.kt
@@ -1,6 +1,0 @@
-package com.recodream_aos.recordream.util.recorder // ktlint-disable package-name
-
-enum class PlayButtonState {
-    RECORDER_PLAY,
-    RECORDER_STOP,
-}
