--- app/src/main/java/com/recodream_aos/recordream/util/recorder/RecordButtonState.kt
@@ -1,7 +1,0 @@
-package com.recodream_aos.recordream.util.recorder // ktlint-disable package-name
-
-enum class RecordButtonState {
-    BEFORE_RECORDING,
-    ON_RECORDING,
-    AFTER_RECORDING,
-}
