+++ app/src/main/res/layout/fragment_record_bottom_sheet.xml
@@ -38,7 +38,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/v_recording_indicator" />
 
-        <com.recodream_aos.recordream.util.recorder.TimeStampTextView
+        <com.recodream_aos.recordream.presentation.record.recording.uistate.TimeStampTextView
             android:id="@+id/tv_recording_progress_time"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -82,7 +82,7 @@
             app:layout_constraintStart_toEndOf="@+id/pb_recording_progress_bar"
             app:layout_constraintTop_toTopOf="@+id/pb_recording_progress_bar" />
 
-        <com.recodream_aos.recordream.util.recorder.RecordButton
+        <com.recodream_aos.recordream.presentation.record.recording.uistate.RecordButton
             android:id="@+id/iv_recording_record_state_btn"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -94,7 +94,7 @@
             app:layout_constraintTop_toBottomOf="@+id/pb_recording_progress_bar"
             tools:background="@drawable/icn_mic_start" />
 
-        <com.recodream_aos.recordream.util.recorder.PlayButton
+        <com.recodream_aos.recordream.presentation.record.recording.uistate.PlayButton
             android:id="@+id/iv_recording_play_state_btn"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
